+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3N1H0Y8PHTPGAXCENNDBBM2"
+        "01K3N7JN4VM642PQRT6W826N3P"
     )
     r"""The id of the resource"""