+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3N1H0Y7ZD5A2GK9RMEREGVG"
+        "01K3N7JN4TN7S2W4B7WBF28VHW"
     )
     r"""The id of the resource"""