+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3N1H0Y8D2XF7SBCAERCEV8K"
+        "01K3N7JN4VCWBDB2TPRDDAZJSE"
     )
     r"""The id of the resource"""